Calculate Coverage Metrics Across Databases

Calculate Coverage Metrics Across Databases

Usage

calc_coverage(results_by_database, gold_standard)

Arguments

results_by_database

List of result sets by database

gold_standard

Vector of relevant article IDs

Value

List containing coverage statistics

Details

Calculates coverage metrics for each database and overall:

coverage_count: Number of relevant articles found by each database

coverage_rate: Proportion of relevant articles found by each database

unique_coverage: Number of relevant articles found only by this database

total_coverage: Overall proportion of relevant articles found by all databases

redundancy_rate: Proportion of duplicate results across databases

Examples

# Create sample data
results_db1 <- c("art1", "art2", "art3", "art4")
results_db2 <- c("art2", "art3", "art5", "art6")
results_by_db <- list("Database1" = results_db1, "Database2" = results_db2)
gold_standard <- c("art1", "art3", "art5", "art7", "art8")

coverage <- calc_coverage(results_by_db, gold_standard)
print(coverage$total_coverage)
#> [1] 0.6